LPS-120225 Refactor
