+++ modules/apps/portal/portal-upgrade-impl/src/main/java/com/liferay/portal/upgrade/internal/release/osgi/commands/ReleaseManagerOSGiCommands.java
@@ -14,47 +14,24 @@
 
 package com.liferay.portal.upgrade.internal.release.osgi.commands;
 
-import com.liferay.osgi.service.tracker.collections.map.PropertyServiceReferenceComparator;
-import com.liferay.osgi.service.tracker.collections.map.PropertyServiceReferenceMapper;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapListener;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.configuration.Configuration;
-import com.liferay.portal.kernel.configuration.ConfigurationFactoryUtil;
-import com.liferay.portal.kernel.dao.db.DB;
-import com.liferay.portal.kernel.dao.db.DBManagerUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.Release;
-import com.liferay.portal.kernel.service.ReleaseLocalService;
-import com.liferay.portal.kernel.upgrade.UpgradeStep;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.output.stream.container.constants.OutputStreamContainerConstants;
 import com.liferay.portal.upgrade.internal.executor.SwappedLogExecutor;
 import com.liferay.portal.upgrade.internal.executor.UpgradeExecutor;
 import com.liferay.portal.upgrade.internal.graph.ReleaseGraphManager;
 import com.liferay.portal.upgrade.internal.registry.UpgradeInfo;
-import com.liferay.portal.upgrade.internal.registry.UpgradeStepRegistratorThreadLocal;
-import com.liferay.portal.util.PropsValues;
+import com.liferay.portal.upgrade.internal.release.ReleaseManager;
 
-import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Properties;
 import java.util.Set;
 
 import org.apache.felix.service.command.Descriptor;
 
-import org.osgi.framework.BundleContext;
-import org.osgi.framework.ServiceReference;
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
-import org.osgi.util.tracker.ServiceTrackerCustomizer;
 
 /**
  * @author Miguel Pastor
@@ -75,14 +52,15 @@ public class ReleaseManagerOSGiCommands {
 	public String check() {
 		StringBundler sb = new StringBundler(0);
 
-		Set<String> bundleSymbolicNames = _serviceTrackerMap.keySet();
+		Set<String> bundleSymbolicNames =
+			_releaseManager.getBundleSymbolicNames();
 
 		for (String bundleSymbolicName : bundleSymbolicNames) {
-			String schemaVersionString = getSchemaVersionString(
+			String schemaVersionString = _releaseManager.getSchemaVersionString(
 				bundleSymbolicName);
 
 			ReleaseGraphManager releaseGraphManager = new ReleaseGraphManager(
-				_serviceTrackerMap.getService(bundleSymbolicName));
+				_releaseManager.getUpgradeInfos(bundleSymbolicName));
 
 			List<List<UpgradeInfo>> upgradeInfosList =
 				releaseGraphManager.getUpgradeInfosList(schemaVersionString);
@@ -128,14 +106,14 @@ public class ReleaseManagerOSGiCommands {
 
 	@Descriptor("Execute upgrade for a specific module")
 	public String execute(String bundleSymbolicName) {
-		if (_serviceTrackerMap.getService(bundleSymbolicName) == null) {
+		List<UpgradeInfo> upgradeInfos = _releaseManager.getUpgradeInfos(
+			bundleSymbolicName);
+
+		if (upgradeInfos == null) {
 			return "No upgrade processes registered for " + bundleSymbolicName;
 		}
 
 		try {
-			List<UpgradeInfo> upgradeInfos = _serviceTrackerMap.getService(
-				bundleSymbolicName);
-
 			_upgradeExecutor.execute(bundleSymbolicName, upgradeInfos, null);
 		}
 		catch (Throwable throwable) {
@@ -153,14 +131,18 @@ public class ReleaseManagerOSGiCommands {
 
 	@Descriptor("Execute upgrade for a specific module and final version")
 	public String execute(String bundleSymbolicName, String toVersionString) {
-		if (_serviceTrackerMap.getService(bundleSymbolicName) == null) {
+		List<UpgradeInfo> upgradeInfos = _releaseManager.getUpgradeInfos(
+			bundleSymbolicName);
+
+		if (upgradeInfos == null) {
 			return "No upgrade processes registered for " + bundleSymbolicName;
 		}
 
-		String schemaVersionString = getSchemaVersionString(bundleSymbolicName);
+		String schemaVersionString = _releaseManager.getSchemaVersionString(
+			bundleSymbolicName);
 
 		ReleaseGraphManager releaseGraphManager = new ReleaseGraphManager(
-			_serviceTrackerMap.getService(bundleSymbolicName));
+			upgradeInfos);
 
 		_upgradeExecutor.executeUpgradeInfos(
 			bundleSymbolicName,
@@ -202,11 +184,12 @@ public class ReleaseManagerOSGiCommands {
 
 	@Descriptor("List registered upgrade processes for all modules")
 	public String list() {
-		Set<String> keySet = _serviceTrackerMap.keySet();
+		Set<String> bundleSymbolicNames =
+			_releaseManager.getBundleSymbolicNames();
 
-		StringBundler sb = new StringBundler(2 * keySet.size());
+		StringBundler sb = new StringBundler(2 * bundleSymbolicNames.size());
 
-		for (String bundleSymbolicName : keySet) {
+		for (String bundleSymbolicName : bundleSymbolicNames) {
 			sb.append(list(bundleSymbolicName));
 			sb.append(StringPool.NEW_LINE);
 		}
@@ -218,18 +201,18 @@ public class ReleaseManagerOSGiCommands {
 
 	@Descriptor("List registered upgrade processes for a specific module")
 	public String list(String bundleSymbolicName) {
-		List<UpgradeInfo> upgradeProcesses = _serviceTrackerMap.getService(
+		List<UpgradeInfo> upgradeInfos = _releaseManager.getUpgradeInfos(
 			bundleSymbolicName);
 
-		StringBundler sb = new StringBundler(5 + (3 * upgradeProcesses.size()));
+		StringBundler sb = new StringBundler(5 + (3 * upgradeInfos.size()));
 
 		sb.append("Registered upgrade processes for ");
 		sb.append(bundleSymbolicName);
 		sb.append(StringPool.SPACE);
-		sb.append(getSchemaVersionString(bundleSymbolicName));
+		sb.append(_releaseManager.getSchemaVersionString(bundleSymbolicName));
 		sb.append(StringPool.NEW_LINE);
 
-		for (UpgradeInfo upgradeProcess : upgradeProcesses) {
+		for (UpgradeInfo upgradeProcess : upgradeInfos) {
 			sb.append(StringPool.TAB);
 			sb.append(upgradeProcess);
 			sb.append(StringPool.NEW_LINE);
@@ -240,60 +223,11 @@ public class ReleaseManagerOSGiCommands {
 		return sb.toString();
 	}
 
-	@Activate
-	protected void activate(final BundleContext bundleContext) {
-		DB db = DBManagerUtil.getDB();
-
-		_serviceTrackerMap = ServiceTrackerMapFactory.openMultiValueMap(
-			bundleContext, UpgradeStep.class,
-			StringBundler.concat(
-				"(&(upgrade.bundle.symbolic.name=*)(|(upgrade.db.type=any)",
-				"(upgrade.db.type=", db.getDBType(), ")))"),
-			new PropertyServiceReferenceMapper<String, UpgradeStep>(
-				"upgrade.bundle.symbolic.name"),
-			new UpgradeServiceTrackerCustomizer(bundleContext),
-			Collections.reverseOrder(
-				new PropertyServiceReferenceComparator<UpgradeStep>(
-					"upgrade.from.schema.version")),
-			new UpgradeInfoServiceTrackerMapListener());
-
-		Set<String> upgradedBundleSymbolicNames = new HashSet<>();
-
-		Set<String> bundleSymbolicNames = _serviceTrackerMap.keySet();
-
-		while (upgradedBundleSymbolicNames.addAll(bundleSymbolicNames)) {
-			for (String bundleSymbolicName : bundleSymbolicNames) {
-				if (!PropsValues.UPGRADE_DATABASE_AUTO_RUN &&
-					(_releaseLocalService.fetchRelease(bundleSymbolicName) !=
-						null)) {
-
-					continue;
-				}
-
-				List<UpgradeInfo> upgradeSteps = _serviceTrackerMap.getService(
-					bundleSymbolicName);
-
-				_upgradeExecutor.execute(
-					bundleSymbolicName, upgradeSteps,
-					OutputStreamContainerConstants.FACTORY_NAME_DUMMY);
-			}
-
-			bundleSymbolicNames = _serviceTrackerMap.keySet();
-		}
-
-		_activated = true;
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		_serviceTrackerMap.close();
-	}
-
 	protected void executeAll(
 		Set<String> upgradeThrewExceptionBundleSymbolicNames) {
 
 		Set<String> upgradableBundleSymbolicNames =
-			getUpgradableBundleSymbolicNames();
+			_releaseManager.getUpgradableBundleSymbolicNames();
 
 		upgradableBundleSymbolicNames.removeAll(
 			upgradeThrewExceptionBundleSymbolicNames);
@@ -306,8 +240,9 @@ public class ReleaseManagerOSGiCommands {
 				upgradableBundleSymbolicNames) {
 
 			try {
-				List<UpgradeInfo> upgradeInfos = _serviceTrackerMap.getService(
-					upgradableBundleSymbolicName);
+				List<UpgradeInfo> upgradeInfos =
+					_releaseManager.getUpgradeInfos(
+						upgradableBundleSymbolicName);
 
 				_upgradeExecutor.execute(
 					upgradableBundleSymbolicName, upgradeInfos, null);
@@ -329,175 +264,16 @@ public class ReleaseManagerOSGiCommands {
 		executeAll(upgradeThrewExceptionBundleSymbolicNames);
 	}
 
-	protected String getSchemaVersionString(String bundleSymbolicName) {
-		String schemaVersionString = "0.0.0";
-
-		Release release = _releaseLocalService.fetchRelease(bundleSymbolicName);
-
-		if ((release != null) &&
-			Validator.isNotNull(release.getSchemaVersion())) {
-
-			schemaVersionString = release.getSchemaVersion();
-		}
-
-		return schemaVersionString;
-	}
-
-	protected Set<String> getUpgradableBundleSymbolicNames() {
-		Set<String> upgradableBundleSymbolicNames = new HashSet<>();
-
-		for (String bundleSymbolicName : _serviceTrackerMap.keySet()) {
-			if (isUpgradable(bundleSymbolicName)) {
-				upgradableBundleSymbolicNames.add(bundleSymbolicName);
-			}
-		}
-
-		return upgradableBundleSymbolicNames;
-	}
-
-	protected boolean isUpgradable(String bundleSymbolicName) {
-		String schemaVersionString = getSchemaVersionString(bundleSymbolicName);
-
-		ReleaseGraphManager releaseGraphManager = new ReleaseGraphManager(
-			_serviceTrackerMap.getService(bundleSymbolicName));
-
-		List<List<UpgradeInfo>> upgradeInfosList =
-			releaseGraphManager.getUpgradeInfosList(schemaVersionString);
-
-		if (upgradeInfosList.size() == 1) {
-			return true;
-		}
-
-		return false;
-	}
-
-	@Reference(unbind = "-")
-	protected void setReleaseLocalService(
-		ReleaseLocalService releaseLocalService) {
-
-		_releaseLocalService = releaseLocalService;
-	}
-
 	private static final Log _log = LogFactoryUtil.getLog(
 		ReleaseManagerOSGiCommands.class);
 
-	private boolean _activated;
-	private ReleaseLocalService _releaseLocalService;
-	private ServiceTrackerMap<String, List<UpgradeInfo>> _serviceTrackerMap;
+	@Reference
+	private ReleaseManager _releaseManager;
 
 	@Reference
 	private SwappedLogExecutor _swappedLogExecutor;
 
 	@Reference
 	private UpgradeExecutor _upgradeExecutor;
-
-	private static class UpgradeServiceTrackerCustomizer
-		implements ServiceTrackerCustomizer<UpgradeStep, UpgradeInfo> {
-
-		public UpgradeServiceTrackerCustomizer(BundleContext bundleContext) {
-			_bundleContext = bundleContext;
-		}
-
-		@Override
-		public UpgradeInfo addingService(
-			ServiceReference<UpgradeStep> serviceReference) {
-
-			String fromSchemaVersionString =
-				(String)serviceReference.getProperty(
-					"upgrade.from.schema.version");
-			String toSchemaVersionString = (String)serviceReference.getProperty(
-				"upgrade.to.schema.version");
-			Object buildNumberObject = serviceReference.getProperty(
-				"build.number");
-
-			UpgradeStep upgradeStep = _bundleContext.getService(
-				serviceReference);
-
-			if (upgradeStep == null) {
-				if (_log.isWarnEnabled()) {
-					_log.warn(
-						"Skipping service " + serviceReference +
-							" because it does not implement UpgradeStep");
-				}
-
-				return null;
-			}
-
-			int buildNumber = 0;
-
-			if (buildNumberObject == null) {
-				try {
-					Class<? extends UpgradeStep> clazz = upgradeStep.getClass();
-
-					Configuration configuration =
-						ConfigurationFactoryUtil.getConfiguration(
-							clazz.getClassLoader(), "service");
-
-					Properties properties = configuration.getProperties();
-
-					buildNumber = GetterUtil.getInteger(
-						properties.getProperty("build.number"));
-				}
-				catch (Exception exception) {
-					if (_log.isDebugEnabled()) {
-						_log.debug(
-							"Unable to read service.properties for " +
-								serviceReference);
-					}
-				}
-			}
-			else {
-				buildNumber = GetterUtil.getInteger(buildNumberObject);
-			}
-
-			return new UpgradeInfo(
-				fromSchemaVersionString, toSchemaVersionString, buildNumber,
-				upgradeStep);
-		}
-
-		@Override
-		public void modifiedService(
-			ServiceReference<UpgradeStep> serviceReference,
-			UpgradeInfo upgradeInfo) {
-		}
-
-		@Override
-		public void removedService(
-			ServiceReference<UpgradeStep> serviceReference,
-			UpgradeInfo upgradeInfo) {
-
-			_bundleContext.ungetService(serviceReference);
-		}
-
-		private final BundleContext _bundleContext;
-
-	}
-
-	private class UpgradeInfoServiceTrackerMapListener
-		implements ServiceTrackerMapListener
-			<String, UpgradeInfo, List<UpgradeInfo>> {
-
-		@Override
-		public void keyEmitted(
-			ServiceTrackerMap<String, List<UpgradeInfo>> serviceTrackerMap,
-			final String key, UpgradeInfo upgradeInfo,
-			List<UpgradeInfo> upgradeInfos) {
-
-			if (_activated && UpgradeStepRegistratorThreadLocal.isEnabled() &&
-				(PropsValues.UPGRADE_DATABASE_AUTO_RUN ||
-				 (_releaseLocalService.fetchRelease(key) == null))) {
-
-				_upgradeExecutor.execute(key, upgradeInfos, null);
-			}
-		}
-
-		@Override
-		public void keyRemoved(
-			ServiceTrackerMap<String, List<UpgradeInfo>> serviceTrackerMap,
-			String key, UpgradeInfo upgradeInfo,
-			List<UpgradeInfo> upgradeInfos) {
-		}
-
-	}
 
 }
